# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/sandbox/runs": {
      "post": {
        "description": "Upload any local file (PDF, DOCX, XLSX, TXT, CSV, HTML, JSON, \u2026) and run one\nor more detectors against its extracted text.\n\n**`detectors`** is a JSON string containing an array of detector config objects.\nEach object has the shape:\n```json\n{ \"type\": \"<TYPE>\", \"enabled\": true, \"config\": { ... } }\n```\n\n### Detector types & sample configs\n\n| Type | What it finds |\n|------|---------------|\n| `SECRETS` | API keys, tokens, private keys |\n| `PII` | Emails, SSNs, credit cards, phone numbers |\n| `TOXIC` | Toxic / hateful text (Detoxify model) |\n| `IMAGE_CLASSIFICATION` | Image content classification (NSFW, etc.) |\n| `YARA` | Custom YARA rule matches |\n| `BROKEN_LINKS` | Unreachable URLs in text |\n\n**Minimal \u2014 secrets only (all patterns):**\n```json\n[{\"type\":\"SECRETS\",\"enabled\":true,\"config\":{}}]\n```\n\n**PII with specific patterns:**\n```json\n[{\"type\":\"PII\",\"enabled\":true,\"config\":{\"enabled_patterns\":[\"email\",\"credit_card\",\"ssn\",\"phone_number\"],\"confidence_threshold\":0.8}}]\n```\n\n**Secrets + PII combined:**\n```json\n[\n  {\"type\":\"SECRETS\",\"enabled\":true,\"config\":{\"enabled_patterns\":[\"aws\",\"github\",\"stripe\",\"generic_api_key\"]}},\n  {\"type\":\"PII\",\"enabled\":true,\"config\":{\"enabled_patterns\":[\"email\",\"ssn\",\"credit_card\"],\"confidence_threshold\":0.75}}\n]\n```\n\n**Toxic content (multilingual model):**\n```json\n[{\"type\":\"TOXIC\",\"enabled\":true,\"config\":{\"enabled_patterns\":[\"toxicity\",\"severe_toxicity\",\"insult\"],\"model_name\":\"multilingual\"}}]\n```\n\n**Full scan \u2014 all detectors:**\n```json\n[\n  {\"type\":\"SECRETS\",\"enabled\":true,\"config\":{}},\n  {\"type\":\"PII\",\"enabled\":true,\"config\":{\"confidence_threshold\":0.7}},\n  {\"type\":\"TOXIC\",\"enabled\":true,\"config\":{}},\n  {\"type\":\"BROKEN_LINKS\",\"enabled\":true,\"config\":{}}\n]\n```",
        "operationId": "SandboxController_createRun",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "file",
                  "detectors"
                ],
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "description": "File to scan"
                  },
                  "detectors": {
                    "type": "string",
                    "description": "JSON array of detector config objects",
                    "example": "[{\"type\":\"SECRETS\",\"enabled\":true,\"config\":{\"enabled_patterns\":[\"aws\",\"github\",\"stripe\",\"generic_api_key\"]}},{\"type\":\"PII\",\"enabled\":true,\"config\":{\"enabled_patterns\":[\"email\",\"credit_card\",\"ssn\",\"phone_number\"],\"confidence_threshold\":0.75}}]"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SandboxRunDto"
                }
              }
            }
          }
        },
        "summary": "Upload a file and run detectors on it",
        "tags": [
          "Sandbox"
        ]
      },
      "get": {
        "operationId": "SandboxController_listRuns",
        "parameters": [
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Search by file name or MIME type",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Filter by one or more run statuses",
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "enum": [
                  "PENDING",
                  "RUNNING",
                  "COMPLETED",
                  "ERROR"
                ]
              }
            }
          },
          {
            "name": "contentType",
            "required": false,
            "in": "query",
            "description": "Filter by one or more content types",
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "enum": [
                  "TXT",
                  "TABLE",
                  "IMAGE",
                  "VIDEO",
                  "AUDIO",
                  "URL",
                  "BINARY",
                  "OTHER"
                ]
              }
            }
          },
          {
            "name": "detectorType",
            "required": false,
            "in": "query",
            "description": "Filter by detectors used in the run configuration or reported findings",
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "enum": [
                  "SECRETS",
                  "PII",
                  "TOXIC",
                  "IMAGE_CLASSIFICATION",
                  "YARA",
                  "BROKEN_LINKS",
                                    "TEXT_CLASSIFICATION",
                  "LANGUAGE",
                  "CODE_SECURITY",
                  "CUSTOM"
                ]
              }
            }
          },
          {
            "name": "hasFindings",
            "required": false,
            "in": "query",
            "description": "Filter by whether findings exist",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "Sort field",
            "schema": {
              "default": "CREATED_AT",
              "type": "string",
              "enum": [
                "CREATED_AT",
                "FILE_NAME",
                "STATUS",
                "FILE_SIZE_BYTES",
                "DURATION_MS",
                "FINDINGS_COUNT"
              ]
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "description": "Sort direction",
            "schema": {
              "default": "DESC",
              "type": "string",
              "enum": [
                "ASC",
                "DESC"
              ]
            }
          },
          {
            "name": "skip",
            "required": false,
            "in": "query",
            "schema": {
              "default": 0,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "default": 50,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SandboxRunListResponseDto"
                }
              }
            }
          }
        },
        "summary": "List sandbox runs (paginated)",
        "tags": [
          "Sandbox"
        ]
      }
    },
    "/sandbox/runs/{id}": {
      "get": {
        "operationId": "SandboxController_getRun",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SandboxRunDto"
                }
              }
            }
          }
        },
        "summary": "Get a sandbox run by ID",
        "tags": [
          "Sandbox"
        ]
      },
      "delete": {
        "description": "Deletes a sandbox run record. If the run is currently in progress the CLI process is killed first.",
        "operationId": "SandboxController_deleteRun",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Run deleted"
          }
        },
        "summary": "Delete a sandbox run",
        "tags": [
          "Sandbox"
        ]
      }
    },
...
  },
  "info": {
    "title": "Classifyre API",
    "description": "Metadata ingestion and detection API for unstructured data sources. Supports WordPress, Slack, S3-Compatible Storage, Azure Blob Storage, Google Cloud Storage, PostgreSQL, MySQL, MSSQL, Oracle, Hive, Databricks, Snowflake, MongoDB, PowerBI, Tableau, Confluence, Jira, and Service Desk sources. Built-in detectors for secrets, PII, toxic content, image classification, broken links, and security threats.",
    "version": "1.0.0",
    "contact": {
      "name": "Classifyre Team",
      "url": "https://github.com/unstructured/classifyre",
      "email": "support@example.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "SandboxRunDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "fileName": {
            "type": "string"
          },
          "fileType": {
            "type": "string",
            "description": "Raw MIME type detected by the CLI (e.g. \"application/pdf\")"
          },
          "contentType": {
            "type": "string",
            "enum": [
              "TXT",
              "TABLE",
              "IMAGE",
              "VIDEO",
              "AUDIO",
              "URL",
              "BINARY",
              "OTHER"
            ],
            "description": "Internal content classification derived from the MIME type"
          },
          "fileExtension": {
            "type": "string"
          },
          "fileSizeBytes": {
            "type": "number"
          },
          "detectors": {
            "type": "object"
          },
          "findings": {
            "type": "object"
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "RUNNING",
              "COMPLETED",
              "ERROR"
            ]
          },
          "errorMessage": {
            "type": "string"
          },
          "durationMs": {
            "type": "number"
          }
        },
        "required": [
          "id",
          "createdAt",
          "fileName",
          "fileType",
          "contentType",
          "fileExtension",
          "fileSizeBytes",
          "detectors",
          "findings",
          "status"
        ]
      },
      "SandboxRunListResponseDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SandboxRunDto"
            }
          },
          "total": {
            "type": "number"
          },
          "skip": {
            "type": "number"
          },
          "limit": {
            "type": "number"
          }
        },
        "required": [
          "items",
          "total",
          "skip",
          "limit"
        ]
      },
...
    }
  }
}